Add Gemfile.lock to .gitignore
Committer Note:
Gemfile is now used by rubygems, but Gemfile.lock file is specific to bundler.
This project does not intend to use bundler, and my opinion is that ignoring
the lock file should be done on individual's environment, not the project.
However, since bundler usage is so common it makes sense to help folks by
ignoring this file for them. Thanks @grosser.